files/requirements.txt
customtkinter
Pillow
easyocr
pdfplumber
opencv-python
numpy
natasha
requests